create dateTimeUtils file, export/import utility functions from there
