+++ lib/Datepicker/Datepicker.js
@@ -13,56 +13,9 @@ import IconButton from '../IconButton';
 import TextField from '../TextField';
 import Calendar from './Calendar';
 import css from './Calendar.css';
+import { getLocaleDateFormat } from '../../util/dateTimeUtils';
 
 const pickDataProps = (props) => pick(props, (v, key) => key.indexOf('data-test') !== -1);
-
-export function getMomentLocalizedFormat(intl) {
-  moment.locale(intl.locale);
-  const format = moment.localeData()._longDateFormat.L;
-  return format;
-}
-
-// Returns a localized format.
-export const getLocaleDateFormat = ({ intl }) => {
-  const tempDate = new Date('Thu May 14 2020 14:39:25 GMT-0500');
-  let format = '';
-
-  // set up a locally formatted array of parts...
-  if (Intl?.DateTimeFormat()?.formatToParts) {
-    const intlFormatter = new Intl.DateTimeFormat(intl.locale, {
-      day: '2-digit',
-      year: 'numeric',
-      month: '2-digit'
-    });
-
-    const formatted = intlFormatter.formatToParts(tempDate);
-
-    // pull the localized format from the array of formatted parts.
-
-    formatted.forEach((p) => {
-      switch (p.type) {
-        case 'month':
-          format += 'MM';
-          break;
-        case 'day':
-          format += 'DD';
-          break;
-        case 'year':
-          format += 'YYYY';
-          break;
-        case 'literal':
-          format += p.value;
-          break;
-        default:
-          break;
-      }
-    });
-  } else {
-    format = getMomentLocalizedFormat(intl);
-  }
-
-  return format;
-};
 
 // Controls the formatting from the value prop to what displays in the UI.
 // need to judge the breakage factor in adopting a spread syntax for these parameters...
